+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Refactor the encryption plugin so no more plugin specific code is in the RxDB core.
 - Refactor local documents plugin to only create a storage instance for local documents when needed.
 - Added `RxDocument().toMutableJSON()`
+- Added `RxCollection().bulkUpsert()`
 
 ### 11.6.0 (4 February 2022)